+++ conf/settings.py
@@ -59,7 +59,6 @@ INSTALLED_APPS = [
     'contact',
     'marketaccess',
     'community',
-    'marketing',
     'search',
     'ukef',
     'healthcheck',
@@ -379,6 +378,14 @@ CONTACT_SOO_ZENDESK_SUBJECT = env.str(
     'CONTACT_DOMESTIC_ZENDESK_SUBJECT',
     'great.gov.uk Selling Online Overseas contact form'
 )
+CONTACT_LOCAL_EXPORT_SUPPORT_AGENT_NOTIFY_TEMPLATE_ID = env.str(
+    'CONTACT_LOCAL_EXPORT_SUPPORT_AGENT_NOTIFY_TEMPLATE_ID',
+    '6e6f8a6b-ff31-490d-8dc5-05225fffaf37'
+)
+CONTACT_LOCAL_EXPORT_SUPPORT_NOTIFY_TEMPLATE_ID = env.str(
+    'CONTACT_LOCAL_EXPORT_SUPPORT_NOTIFY_TEMPLATE_ID',
+    '5d692991-0c57-4ae0-9053-95fd8ddf53eb'
+)
 CONTACT_EVENTS_USER_NOTIFY_TEMPLATE_ID = env.str(
     'CONTACT_EVENTS_USER_NOTIFY_TEMPLATE_ID',
     '2d5d556a-e0fa-4a9b-81a0-6ed3fcb2e3da'
@@ -397,6 +404,16 @@ CONTACT_DSO_USER_NOTIFY_TEMPLATE_ID = env.str(
     'CONTACT_DSO_USER_NOTIFY_TEMPLATE_ID',
     'a6a3db79-944f-4c59-8eeb-2f756019976c'
 )
+CONTACT_ECOMMERCE_EXPORT_SUPPORT_AGENT_EMAIL_ADDRESS = env.str('CONTACT_ECOMMERCE_EXPORT_SUPPORT_AGENT_EMAIL_ADDRESS')
+CONTACT_ECOMMERCE_EXPORT_SUPPORT_AGENT_NOTIFY_TEMPLATE_ID = env.str(
+    'CONTACT_ECOMMERCE_EXPORT_SUPPORT_AGENT_NOTIFY_TEMPLATE_ID',
+    'a56114d3-515e-4ee7-bb1a-9a0ceab04378'
+)
+CONTACT_ECOMMERCE_EXPORT_SUPPORT_NOTIFY_TEMPLATE_ID = env.str(
+    'CONTACT_ECOMMERCE_EXPORT_SUPPORT_NOTIFY_TEMPLATE_ID',
+    '18d807d2-f4cf-4b93-96c1-0d3169bd0906',
+)
+
 CONTACT_DIT_AGENT_EMAIL_ADDRESS = env.str('CONTACT_DIT_AGENT_EMAIL_ADDRESS')
 CONTACT_INTERNATIONAL_AGENT_NOTIFY_TEMPLATE_ID = env.str(
     'CONTACT_INTERNATIONAL_AGENT_NOTIFY_TEMPLATE_ID',
