+++ core/templates/core/request-export-support-form.html
@@ -24,7 +24,7 @@
 		<div class="container">
 			<div class="column-full">
 				<div class="exred-triage-form">
-					<form action="{% url 'marketing-join-form' %}" method="post" class="{% block form_class %}lg-6{% endblock %}">
+					<form method="post" class="{% block form_class %}lg-6{% endblock %}">
 						{% csrf_token %}
 
 						<fieldset class="margin-bottom-20">
