--- marketing/constants.py
@@ -1,7 +1,0 @@
-EMPLOYEES_NUMBER_CHOICES = (
-    ('1-9', '1 to 9'),
-    ('10-49', '10 to 49'),
-    ('50-249', '50 to 249'),
-    ('250-499', '250 to 499'),
-    ('500plus', 'More than 500'),
-)
